requirements.txt
mcp[cli]
psycopg2-binary
python-dotenv
uvicorn
uv
httpx
asyncpg
